(feat) Executing on stage macro when content is created via scheduled template.
